refactor: In Domain's model.game's groups -> model.game.group.
