+++ app/src/main/java/no/usn/mob3000/data/repository/game/GroupsRepository.kt
@@ -3,12 +3,11 @@ package no.usn.mob3000.data.repository.game
 import android.util.Log
 import io.github.jan.supabase.SupabaseClient
 import io.github.jan.supabase.postgrest.postgrest
-import no.usn.mob3000.data.model.game.OpeningsDto
 import no.usn.mob3000.data.model.game.RepertoireDto
 import no.usn.mob3000.data.network.SupabaseClientWrapper
 import no.usn.mob3000.data.source.remote.game.GroupsDataSource
 import no.usn.mob3000.domain.helper.game.mapToDomain
-import no.usn.mob3000.domain.model.game.Group
+import no.usn.mob3000.domain.model.game.group.Group
 import no.usn.mob3000.domain.repository.game.IGroupsRepository
 
 /**
