+++ app/src/main/java/no/usn/mob3000/domain/usecase/game/group/UpdateGroupUseCase.kt
@@ -1,7 +1,7 @@
 package no.usn.mob3000.domain.usecase.game.group
 
 import no.usn.mob3000.domain.helper.game.mapToData
-import no.usn.mob3000.domain.model.game.Group
+import no.usn.mob3000.domain.model.game.group.Group
 import no.usn.mob3000.domain.repository.game.IGroupsRepository
 
 /**
