+++ app/src/main/java/no/usn/mob3000/ui/screens/chess/train/group/UpdateGroupScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.res.stringResource
 import no.usn.mob3000.R
-import no.usn.mob3000.domain.model.game.Group
+import no.usn.mob3000.domain.model.game.group.Group
 import no.usn.mob3000.domain.model.game.Opening
 import no.usn.mob3000.ui.components.game.group.GroupEditor
